+++ internal/zero/telemetry/analytics/sessions.go
