+++ internal/zero/telemetry/analytics/storage.go
