+++ internal/zero/telemetry/analytics/storage_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/pomerium/pomerium/internal/zero/analytics"
+	"github.com/pomerium/pomerium/internal/zero/telemetry/analytics"
 )
 
 func TestStorage(t *testing.T) {
