+++ internal/zero/telemetry/prometheus/stream.go
@@ -1,4 +1,4 @@
-package convert
+package prometheus
 
 import (
 	"bufio"
