+++ internal/zero/telemetry/prometheus/testdata/counter.txt
