+++ internal/zero/telemetry/prometheus/testdata/histogram.txt
